refactor(handshake): move auth logic to a service
